+++ src/actions/create_transfer.ts
@@ -41,6 +41,23 @@ export const create_transfer = async (event: EVENT, options: Config) => {
     } catch (e) {}
 
     const repzo_serial_number = body?.serial_number?.formatted;
+    try {
+      if (body?._id) {
+        body.integration_meta = body?.integration_meta || {};
+        body.integration_meta.sync_to_sap_started = true;
+        body.integration_meta.sync_to_sap_succeeded = false;
+        await repzo.transfer.update(body._id, {
+          integration_meta: body.integration_meta,
+        });
+      }
+    } catch (e) {
+      console.error(e);
+      await actionLog
+        .addDetail(
+          `Failed updating integration_meta of Transfer: ${repzo_serial_number}`
+        )
+        .commit();
+    }
 
     await actionLog
       .addDetail(`Transfer - ${repzo_serial_number} => ${body?.sync_id}`)
@@ -157,6 +174,22 @@ export const create_transfer = async (event: EVENT, options: Config) => {
     );
 
     // console.log(result);
+    try {
+      if (body?._id) {
+        body.integration_meta = body?.integration_meta || {};
+        body.integration_meta.sync_to_sap_succeeded = true;
+        await repzo.transfer.update(body._id, {
+          integration_meta: body.integration_meta,
+        });
+      }
+    } catch (e) {
+      console.error(e);
+      await actionLog
+        .addDetail(
+          `Failed updating integration_meta of Transfer: ${repzo_serial_number}`
+        )
+        .commit();
+    }
 
     await actionLog
       .addDetail(`SAP Responded with `, result)